Resolves #20 

Perf comparisons on the slack app:

| Task | Before | After |
|-|--------|-------|
| inferExternalStagingAndroidTestKeepRulesForKeeper | 14.126s | 11.138s |
| jarExternalStagingClassesForKeeper | 10.364s | 6.347s |
| jarExternalStagingAndroidTestClassesForKeeper | 4.504s | 2.609s |

~~Will do some perf comparisons in our app and report the findings here~~

~~**On hold as ZipFlinger fails explosively when there are colliding names:** https://issuetracker.google.com/issues/150320652~~