@fabiensanglard 2020-03-04
Did you get a chance to gather metrics on performance improvements?