@ZacSweers 2020-03-04
@fabiensanglard not yet as we're not able to fully use it in our app until https://issuetracker.google.com/issues/150320652 is resolved. We did anecdotally see ~50% improvement in APK packaging time in our app after updating to AGP 3.6, which I understand also uses zipflinger now.