@fabiensanglard 2020-03-04
Can you use the suggested workaround for name collision (delete before add)? (the bug is marked as Intended Behavior ;P).