@ZacSweers 2020-03-04
Ah didn't see they'd updated it. Will try it out, though not thrilled that they're sticking with `selectAll()` without having an option for this considering it's a common case. Makes that API a footgun IMO