@ZacSweers 2020-03-04
I'm going to check one more thing WRT to jar contents. I suspect we may be packaging substantially more that the native `Jar` task is better at stripping