@ZacSweers 2020-03-04
Ok updated again, I think I just had some extra work happening on my machine at the same time and now it's faster. I'm not able to get a reasonable diff comparison due to entries being in different order though. Will investigate that as needed separately