@fabiensanglard 2020-03-05
One more thing. If you really want to reduce packing wall-time to the max, it will help a lot to multi-thread building BytesSource and use a Synchronized archive to add them.