@ZacSweers 2020-03-05
Yeah I've thought about that, but I'm hesitant to add concurrency to something that already just runs on release builds. Our app is nontrivial and 6 seconds seems fairly amortized by the cost of R8, building the full app, dexing, etc happening. I'll prototype it just to see what the numbers would look like though!